Added some debug prints.

+++ code/train_val_test_utilities.py
@@ -83,7 +83,7 @@ def run_model(model, dataloader, mode, device, optimizer=None, class_specific=Tr
             print(labels.type(), labels)
 
             # We first compute the CrossEntropy Loss
-            cross_entropy = torch.nn.functional.cross_entropy(logits, labels)
+            cross_entropy = torch.nn.functional.cross_entropy(logits, labels.long())
 
             # Check this condition
             if class_specific:
